+++ docusaurus.config.ts
@@ -100,6 +100,7 @@ const config: Config = {
           label: '文档',
         },
         { to: '/blog', label: '博客', position: 'left'},
+        { to: '/tools', label: '工具', position: 'left'},
         {
           type: 'localeDropdown',
           position: 'right',